% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pps_matrix.R
\name{pps_matrix}
\alias{pps_matrix}
\title{Generate Predictive Power Score for all variable in a data.frame}
\usage{
pps_matrix(df)
}
\arguments{
\item{df}{A data.frame the data to be analyzed}
}
\value{
A data.frame containing 1 row for each PPS calculation:
\itemize{
\item The Target Variable
\item The Feature Variable
\item The Task Name
\item The Predictive Power Score
\item The Evaluation metric
\item The Baseline Score (Evaluate Metric vs. a Zero-R Classifer/Median)
\item The Model Score
}
}
\description{
Calculate the Predictive Power Score (PPS) for all columns in the dataframe.  Data is returned in long-format
}
\examples{
pps_matrix(titanic)
}
\keyword{zuh}
